Implementacion a TODO el proyecto de Bases de Datos con MySQL (Mala decision debi hacer mas commits), Implementacion paquetes de infraestructura

+++ src/main/java/ProyectoPropio1/dto/DatosVentaProductoDTO.java
@@ -1,10 +1,12 @@
 package ProyectoPropio1.dto;
 
+import java.math.BigDecimal;
+
 public interface DatosVentaProductoDTO {
 
     String nombre();
 
-    double precio();
+    BigDecimal precio();
 
 }
