+++ src/main/java/ProyectoPropio1/dto/DatosVentaProductoPerecederoDTO.java
@@ -1,15 +1,16 @@
 package ProyectoPropio1.dto;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
-public record DatosVentaProductoPerecederoDTO(String nombre, double precio, LocalDate fechaVencimiento) implements DatosVentaProductoDTO {
+public record DatosVentaProductoPerecederoDTO(String nombre, BigDecimal precio, LocalDate fechaVencimiento) implements DatosVentaProductoDTO {
     @Override
     public String nombre() {
         return nombre;
     }
 
     @Override
-    public double precio() {
+    public BigDecimal precio() {
         return precio;
     }
 }
