+++ src/main/java/ProyectoPropio1/dto/HistorialVentasDTO.java
@@ -1,7 +1,8 @@
 package ProyectoPropio1.dto;
 
+import java.math.BigDecimal;
 import java.util.List;
 
-public record HistorialVentasDTO(List<FacturaDTO> facturasRegistradas, double recaudoTotal) {
+public record HistorialVentasDTO(List<FacturaDTO> facturasRegistradas, BigDecimal recaudoTotal) {
 }
